Updates pkgwrap and adds new npm lock file
- Adds a new Title component.  This breaks up the Item component into a
reusable Title and Item.
- Adds onNew to the AccordionItem
- Removed key from BaseProps.  This was generating a warning because React
doesn't want the components to touch key after it is "set".  Key should
just be generated and forgotten.
- Added some tests for ech "getDefault..." props helper.
- Removed yarn.lock for new npm 5.0
- Adds a new base module that holds sharable functions.  it holds the
baseClasses function that is used to add general CSS selectors that
every copmonent shares.
- Removed all left/right title for title/widget naming.  The title
control uses a "title" and a custom "wiget" control (which could be an
additional text string).