Changes the way Keys are testing within TabContainer
The TabContainer generates unique keys per instance with the util.keys
module.  The problem is with each new test the UUID keys are different.
A new testing mode was added to the Keys module to allow testing keys to
be generated that will be deterministic for unit testing purposes for
jest snapshot testing.