Fixes all testing snapshots
Changed the way the "testing" prop was set.  It now retrieves the NODE_ENV
variable and uses that to determine if the code is under test.  There is no need
to explicitly pass "testing" to each component under test.  It now automatically
sets every component to "testing=true" when run under jest.