DynamicList now accepts new props for list items
The old DynamicList would not accept new props for list items after initial creation.  This is a problem when working with optional widgets that are given to the control and for items that are created after initial creation.  There was no way to inject the a widget.  The items prop now works like a more traditional react control.